Update to save create templates state.
