+++ src/app/app.ts
@@ -1,6 +1,6 @@
 import { App } from "./app.interface";
 import { Cmd, Cmds } from "./cmds/cmd";
-import { CreateTemplatesCmd } from "./cmds/create_templates.cmd";
+import { CreateTemplatesCmd } from "./cmds/create_templates/create_templates.cmd";
 import { DeleteTranslationCacheCmd } from "./cmds/delete_translation_cache.cmd";
 import { TranslateCmd } from "./cmds/translate/translate.cmd";
 import { ConfigRepositoryImpl } from "./components/config/config.repository";
@@ -82,8 +82,9 @@ export class FigmaTranslator implements App {
     ),
     [Cmds.createTemplates]: new CreateTemplatesCmd(
       this.figmaService,
-      this.templateService,
-      this.platformService
+      this.configService,
+      this.platformService,
+      this.templateService
     ),
     [Cmds.deleteTranslationCache]: new DeleteTranslationCacheCmd(
       this.translatorCacheService
