+++ src/app/cmds/create_templates/create_templates.state.ts
@@ -1,0 +1,24 @@
+import { PlatformLocale } from "../../components/platform/platform.interface";
+import { Template } from "../../components/template/template.interface";
+import { TextDirection } from "../../components/translator_language/translator_language.interface";
+import { Platform } from "../../params/platform.param";
+
+export type CreateTemplatesState = {
+  platform: Platform;
+  textDirection: TextDirection;
+  templateScale: number;
+  targetLocales: {
+    [textDirection in TextDirection]: {
+      targetLocale: PlatformLocale;
+      isChecked: true;
+      isVisible: true;
+    }[];
+  };
+  templates: {
+    [platform in Platform]: {
+      template: Template;
+      isChecked: boolean;
+      count: number;
+    }[];
+  };
+};
